+++ src/yajra/Datatables/Engines/BaseEngine.php
@@ -15,14 +15,14 @@ use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Str;
 use League\Fractal\Manager;
 use League\Fractal\Resource\Collection;
-use League\Fractal\TransformerAbstract;
+use League\Fractal\Serializer\DataArraySerializer;
+use League\Fractal\Serializer\SerializerAbstract;
 use yajra\Datatables\Contracts\DataTableEngine;
 use yajra\Datatables\Helper;
 use yajra\Datatables\Processors\DataProcessor;
 
 abstract class BaseEngine implements DataTableEngine
 {
-
     /**
      * Datatables Request object.
      *
@@ -79,7 +79,6 @@ abstract class BaseEngine implements DataTableEngine
      */
     protected $query_type;
 
-
     /**
      * Extra/Added columns.
      *
@@ -109,21 +108,21 @@ abstract class BaseEngine implements DataTableEngine
     protected $autoFilter = true;
 
     /**
-     * Callback to override global search
+     * Callback to override global search.
      *
      * @var \Closure
      */
     protected $filterCallback;
 
     /**
-     * Parameters to passed on filterCallback
+     * Parameters to passed on filterCallback.
      *
      * @var mixed
      */
     protected $filterCallbackParameters;
 
     /**
-     * DT row templates container
+     * DT row templates container.
      *
      * @var array
      */
@@ -149,7 +148,7 @@ abstract class BaseEngine implements DataTableEngine
     protected $prefix;
 
     /**
-     * Database driver used
+     * Database driver used.
      *
      * @var string
      */
@@ -161,6 +160,13 @@ abstract class BaseEngine implements DataTableEngine
      * @var boolean
      */
     protected $isFilterApplied = false;
+
+    /**
+     * Fractal serializer.
+     *
+     * @var SerializerAbstract
+     */
+    protected $serializer;
 
     /**
      * Setup search keyword.
@@ -546,6 +552,19 @@ abstract class BaseEngine implements DataTableEngine
     }
 
     /**
+     * Set fractal serializer class.
+     *
+     * @param string $serializer
+     * @return $this
+     */
+    public function setSerializer($serializer)
+    {
+        $this->serializer = $serializer;
+
+        return $this;
+    }
+
+    /**
      * Organizes works.
      *
      * @param bool $mDataSupport
@@ -664,6 +683,9 @@ abstract class BaseEngine implements DataTableEngine
             if ($this->request->get('include')) {
                 $fractal->parseIncludes($this->request->get('include'));
             }
+
+            $serializer = $this->serializer ?: Config::get('datatables.fractal.serializer', DataArraySerializer::class);
+            $fractal->setSerializer(new $serializer);
 
             //Get transformer reflection
             //Firs method parameter should be data/object to transform